Delete old router implementation
